backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-socketio==5.10.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.1
langchain==0.1.0
langchain-google-genai==0.0.5
langgraph==0.0.20
httpx==0.25.2
aioredis==2.0.1
celery==5.3.4
python-dotenv==1.0.0
